Edit it the files structure  for Next-intl
