+++ app/[locale]/(main)/courses/page.jsx
