+++ app/[locale]/(main)/home/page.jsx
