+++ app/[locale]/(main)/search/page.jsx
