+++ app/[locale]/layout.js
@@ -1,0 +1,35 @@
+import { NextIntlClientProvider } from 'next-intl';
+import './globals.css'
+import { Inter } from 'next/font/google'
+
+
+const inter = Inter({ subsets: ['latin'] })
+
+export const metadata = {
+  title: 'LearnU Team-5',
+  description: 'Generated by create next app',
+}
+
+export function generateStaticParams() {
+  return [{ locale: 'en' }, { locale: 'tr' }];
+}
+
+export default async function RootLayout({ children, params: { locale } }) {
+
+  let messages;
+  try {
+    messages = (await import(`../../messages/${locale}.json`)).default;
+  } catch (error) {
+    console.log(error);
+  }
+
+  return (
+    <html lang={locale}>
+      <body className={`${inter.className} bg-[#f5f5f5]`}>
+        <NextIntlClientProvider locale={locale} messages={messages}>
+          {children}
+        </NextIntlClientProvider>
+      </body>
+    </html>
+  );
+}
