+++ app/[locale]/page.js
