--- app/layout.js
@@ -1,19 +1,0 @@
-import './globals.css'
-import { Inter } from 'next/font/google'
-
-const inter = Inter({ subsets: ['latin'] })
-
-export const metadata = {
-  title: 'LearnU Team-5',
-  description: 'Generated by create next app',
-}
-
-export default function RootLayout({ children, params: { locale } }) {
-  return (
-    <html lang="en">
-      <body className={`${inter.className} bg-[#f5f5f5]`}>
-        {children}
-        </body>
-    </html>
-  )
-}
