+++ package.json
@@ -20,6 +20,7 @@
     "eslint-config-next": "13.4.7",
     "firebase": "^10.0.0",
     "next": "13.4.7",
+    "next-intl": "^2.19.0",
     "postcss": "8.4.24",
     "react": "18.2.0",
     "react-dom": "18.2.0",
